moved files, updated imports and fixed npm run client command
