+++ Backend/src/_helpers/auth-header.js
@@ -1,4 +1,4 @@
-import { authenticationService } from '@/_services';
+import { authenticationService } from '../_services';
 
 export function authHeader() {
     // return authorization header with jwt token
