+++ Backend/src/_helpers/fake-backend.js
@@ -1,5 +1,5 @@
 import { Role } from '.'
-const apiUrl = import.meta.env.VITE_API_URL;
+const apiUrl = process.env.VITE_API_URL;
 
 export async function configureFakeBackend() {
     // let users = [
