+++ Frontend/src/main.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import ReactDOM from 'react-dom/client'
 import App from './App.jsx'
 
-import { configureFakeBackend } from './_helpers/index.js';
+import { configureFakeBackend } from '../../Backend/src/_helpers/index.js';
 
 configureFakeBackend();
